fix merge + modal + clean
