+++ apps/web/src/app/constants/data.ts
@@ -1,0 +1,9 @@
+import { Contact } from "../../types";
+
+export const CONTACTS_DATA: Contact[] = [
+    { handle: "@gohan", avatarUrl: "/avatar/gohan.jpg" },
+    { handle: "@vegeta", avatarUrl: "/avatar/vegeta.jpeg" },
+    { handle: "@frieza", avatarUrl: "/avatar/frieza.png" },
+    { handle: "@piccolo", avatarUrl: "/avatar/piccolo.jpg" },
+    { handle: "@cell", avatarUrl: "/avatar/cell.jpg" },
+  ];
